Split double-conversion.
