+++ CMakeLists.txt
@@ -6,9 +6,11 @@ set(headers
     double-conversion/cached-powers.h
     double-conversion/diy-fp.h
     double-conversion/double-conversion.h
+    double-conversion/double-to-string.h
     double-conversion/fast-dtoa.h
     double-conversion/fixed-dtoa.h
     double-conversion/ieee.h
+    double-conversion/string-to-double.h
     double-conversion/strtod.h
     double-conversion/utils.h)
 
@@ -17,9 +19,10 @@ add_library(double-conversion
             double-conversion/bignum-dtoa.cc
             double-conversion/cached-powers.cc
             double-conversion/diy-fp.cc
-            double-conversion/double-conversion.cc
+            double-conversion/double-to-string.cc
             double-conversion/fast-dtoa.cc
             double-conversion/fixed-dtoa.cc
+            double-conversion/string-to-double.cc
             double-conversion/strtod.cc
             ${headers})
 target_include_directories(
