+++ double-conversion/SConscript
@@ -4,9 +4,10 @@ double_conversion_sources = [
     'bignum-dtoa.cc',
     'cached-powers.cc',
     'diy-fp.cc',
-    'double-conversion.cc',
+    'double-to-string.cc',
     'fast-dtoa.cc',
     'fixed-dtoa.cc',
+    'string-to-double.cc',
     'strtod.cc'
   ]
 Return('double_conversion_sources')
